Adds shared setMonthAndDate function
